gastromanager/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
backports.zoneinfo==0.2.1
black==23.9.1
boto3==1.28.72
botocore==1.31.72
Django==4.2.6
django-activity-stream==2.0.0
django-storages==1.14.2
django-qrcode==0.3
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==6.1.0
Pillow==10.1.0
platformdirs==3.10.0
pluggy==1.3.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
pypng==0.20220715.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.7.0
six==1.16.0
pyzbar==0.1.9
qrcode==7.4.2
